+++ test/tests/stress.cc
@@ -4,6 +4,7 @@
 
 #include <nobind.h>
 
+#include <memory>
 #include <vector>
 
 constexpr auto NestedAsync = Nobind::ReturnNested | Nobind::ReturnAsync;
@@ -46,6 +47,9 @@ std::vector<Hello *> take_and_return_ptr_vector(const std::vector<Hello *> &inpu
   return output;
 }
 
+std::shared_ptr<Hello> take_and_return_shared_ptr(const std::shared_ptr<Hello> in) { return in; }
+std::shared_ptr<Hello> make_shared_ptr(std::string name) { return std::make_shared<Hello>(name); }
+
 NOBIND_MODULE(stress, m) {
   m.def<Hello>("Hello")
       .cons<std::string &>()
@@ -71,4 +75,7 @@ NOBIND_MODULE(stress, m) {
   // Without the object store this function is quite dangerous
   // (returned pointers are owned by default)
   m.def<&take_and_return_ptr_vector, SharedAsync>("take_and_return_ptr_vector");
+
+  m.def<&take_and_return_shared_ptr, Nobind::ReturnAsync>("take_and_return_shared_ptr");
+  m.def<&make_shared_ptr, Nobind::ReturnAsync>("make_shared_ptr");
 }